+++ lib/domain/menstruation/menstruation_store.dart
@@ -11,7 +11,7 @@ import 'package:pilll/database/diary.dart';
 import 'package:pilll/database/menstruation.dart';
 import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
-import 'package:pilll/service/setting.dart';
+import 'package:pilll/database/setting.dart';
 import 'package:pilll/domain/menstruation/menstruation_state.codegen.dart';
 import 'package:pilll/service/user.dart';
 import 'package:pilll/util/datetime/day.dart';
