+++ lib/domain/record/components/add_pill_sheet_group/add_pill_sheet_group_store.dart
@@ -10,7 +10,7 @@ import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
-import 'package:pilll/service/setting.dart';
+import 'package:pilll/database/setting.dart';
 import 'package:pilll/util/datetime/day.dart';
 import 'package:riverpod/riverpod.dart';
