+++ lib/domain/settings/menstruation/setting_menstruation_store.dart
@@ -1,7 +1,7 @@
 import 'package:pilll/domain/settings/menstruation/setting_menstruation_state.codegen.dart';
 import 'package:pilll/entity/pill_sheet_type.dart';
 import 'package:pilll/entity/setting.codegen.dart';
-import 'package:pilll/service/setting.dart';
+import 'package:pilll/database/setting.dart';
 import 'package:riverpod/riverpod.dart';
 
 final settingMenstruationStoreProvider = StateNotifierProvider.autoDispose<
