+++ test/helper/mock.dart
@@ -10,7 +10,7 @@ import 'package:pilll/database/menstruation.dart';
 import 'package:pilll/database/pill_sheet.dart';
 import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/database/pill_sheet_modified_history.dart';
-import 'package:pilll/service/setting.dart';
+import 'package:pilll/database/setting.dart';
 import 'package:pilll/service/day.dart';
 import 'package:pilll/service/user.dart';
 import 'package:mockito/annotations.dart';
